+++ python/enlighten/enlighten/model/rgm_stats_intervals1.py
@@ -74,7 +74,7 @@ class RgmStatsIntervals1(ModelNormal):
         return {
             'channel': (int,),  # noqa: E501
             'end_at': (int,),  # noqa: E501
-            'wh_del': (int,),  # noqa: E501
+            'wh_del': (float,),  # noqa: E501
             'curr_w': (int,),  # noqa: E501
         }
 
@@ -108,7 +108,7 @@ class RgmStatsIntervals1(ModelNormal):
         Args:
             channel (int): The reporting channel identifier.
             end_at (int): End of interval. The format is Unix epoch time unless you pass a `datetime_format` parameter as described [here](https://developer.enphase.com/docs#Datetimes).
-            wh_del (int): Energy delivered during this interval, in Watt-hours.
+            wh_del (float): Energy delivered during this interval, in Watt-hours.
             curr_w (int): Energy delivered during this interval, in Watts.
 
         Keyword Args:
